feat(api+web): Fixing requests = changing survey + merging sendgrid_mail
