+++ api/api/serializers/survey.py
@@ -5,4 +5,4 @@ from api.models.survey import SurveyResponse
 class SurveyResponseSerializer(serializers.ModelSerializer):
     class Meta:
         model = SurveyResponse
-        fields = ['occupation', 'rating', 'experience', 'source', 'feedback']
+        fields = '_all_'
